First Attemt on creating a DLL Scripting System
Created the Dll project.

Created a component script. Prepared all methods for the component script to run.

CRITICAL FAILURE: Tryed to solve alll the problems with linking the librarys but runned into and enclosing pint. Certain errors in the linkage of the librarys and couldn't build the dll lib to be added on the main project. Run out of time.