Second Attemt on creating a DLL Scripting System
Tried a new aproach for the DLL to work by adding a SDK project to manage the scripts directly and serve as an in between of the engine and the scripts. Taken the reference code of the "MyGameMaker" project.

CRITICAL FAILURE: Linking the projects again was a failure. Missing linkage on aditional librarys. DLL build failure. Run out of time.